requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.2
django-cors-headers==3.5.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
mccabe==0.6.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
